# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jingui API",
    "version": "0.1.1",
    "description": "Current jingui-server API surface (implementation-aligned)"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/v1/credentials/callback": {
      "get": { "summary": "OAuth callback", "responses": { "200": { "description": "Authorized" }, "400": { "description": "Invalid callback" } } }
    },
...
  }
}